requirements.txt
streamlit 
joblib==1.3.2
openai==0.28
tabula-py==2.2.0
scikit-learn==1.2.2
tensorflow==2.15.0
matplotlib==3.8.3
seaborn==0.13.2